Predicting the likelihood of e-signing a loan
Product: A loan

Goal: Develop a model to predict for 'quality' applicants. In this case study, 'quality' applicants are those who reach a key part of the loan application process.

Here, the goal is to create a model that predicts whether or not the applicants will complete the electronic signature phase of the loan application (e_signed).